requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.6.5
Django==5.0.4
idna==3.7
isodate==0.6.1
marshmallow==3.21.2
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
python-dotenv==1.0.1
python-youtube==0.9.4
pytube==15.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.5.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
